:memo: Separated ContributorsScreenRobot from ContributorsScreenTest.
