+++ feature/contributors/src/androidUnitTest/kotlin/io/github/droidkaigi/confsched/contributors/ContributorsScreenTest.kt
@@ -1,17 +1,12 @@
 package io.github.droidkaigi.confsched.contributors
 
-import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.hasTestTag
-import androidx.compose.ui.test.hasText
 import dagger.hilt.android.testing.BindValue
 import dagger.hilt.android.testing.HiltAndroidTest
 import io.github.droidkaigi.confsched.testing.DescribedBehavior
 import io.github.droidkaigi.confsched.testing.describeBehaviors
 import io.github.droidkaigi.confsched.testing.execute
+import io.github.droidkaigi.confsched.testing.robot.ContributorsScreenRobot
 import io.github.droidkaigi.confsched.testing.robot.ContributorsServerRobot
-import io.github.droidkaigi.confsched.testing.robot.DefaultContributorsServerRobot
-import io.github.droidkaigi.confsched.testing.robot.DefaultScreenRobot
-import io.github.droidkaigi.confsched.testing.robot.ScreenRobot
 import io.github.droidkaigi.confsched.testing.robot.runRobot
 import io.github.droidkaigi.confsched.testing.rules.RobotTestRule
 import org.junit.Rule
@@ -66,34 +61,5 @@ class ContributorsScreenTest(private val testCase: DescribedBehavior<Contributor
                 }
             }
         }
-    }
-}
-
-class ContributorsScreenRobot @Inject constructor(
-    screenRobot: DefaultScreenRobot,
-    contributorsServerRobot: DefaultContributorsServerRobot,
-) : ScreenRobot by screenRobot,
-    ContributorsServerRobot by contributorsServerRobot {
-    fun setupScreenContent() {
-        robotTestRule.setContent {
-            ContributorsScreen(
-                onNavigationIconClick = { },
-                onContributorsItemClick = { },
-            )
-        }
-    }
-
-    fun checkContributorsDisplayed() {
-        composeTestRule
-            .onNode(hasTestTag(ContributorsScreenTestTag))
-            .assertIsDisplayed()
-    }
-
-    fun checkErrorSnackbarDisplayed() {
-        composeTestRule
-            .onNode(
-                hasText("Fake IO Exception"),
-                useUnmergedTree = true,
-            ).assertIsDisplayed()
     }
 }
